Mouse-picking merge to main

+++ Aurora-Editor/src/Panels/SceneHierarchyPanel.cpp
@@ -8,6 +8,7 @@ namespace Aurora {
 		:m_scene(scene)
 	{
 	}
+
 	SceneHierarchyPanel::~SceneHierarchyPanel()
 	{
 	}
@@ -41,7 +42,7 @@ namespace Aurora {
 
 		ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_OpenOnArrow | ImGuiTreeNodeFlags_SpanAvailWidth;
 
-		bool opened = ImGui::TreeNodeEx((void*)(uint32_t)*entity, flags, tag.c_str());
+		bool opened = ImGui::TreeNodeEx((void*)(uint64_t)(uint32_t)*entity, flags, tag.c_str());
 
 		if (ImGui::IsItemClicked())
 		{
